requirements.txt
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
eAwaiter @ git+https://github.com/5KRC1/eAwaiter@c5d9d0040ddace2e30e19dc9c8cb367692faf3c4
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
requests==2.28.2
soupsieve==2.4
urllib3==1.26.14
Werkzeug==2.2.2
